requirements.txt
pandas==0.24.2
scikit-learn==0.23.2
numpy==1.18.5
json
requests
python-decouple
